requirements.txt
Django==3.0.4
psycopg2>=2.7,<3.0
psycopg2-binary==2.7.5
django-oauth-toolkit
django-cors-middleware
django-filter==2.2.0
django-oauth-toolkit==1.3.2
django-rest-enumfield==0.2.0
djangorestframework==3.11.0


